third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-events-expected.txt

Tests that RTCPeerConnection event callbacks are async so that for example close can be called safely. The order of the messages is very important.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gotStream was called.
PASS gotStream done.
PASS onNegotiationNeeded was called.
PASS pc.signalingState is 'closed'
PASS onNegotiationNeeded done.
PASS successfullyParsed is true

TEST COMPLETE
